+++ spec/backspin/credential_scrubbing_spec.rb
@@ -12,6 +12,10 @@ RSpec.describe "Backspin credential scrubbing" do
       expect(Backspin.configuration.scrub_credentials).to be true
     end
 
+    it "uses a debug logger by default" do
+      expect(Backspin.configuration.logger.level).to eq(Logger::DEBUG)
+    end
+
     it "can disable credential scrubbing" do
       Backspin.configure do |config|
         config.scrub_credentials = false